misc fix for compiler build: 1: support compiler build at MacOS 2: support build with clang
GitOrigin-RevId: 28abef8a461c83630e768e4a7d75d1a001976b99

+++ compiler/lib/Conversion/MGBToKernel/MGBToKernel.cpp
@@ -566,7 +566,8 @@ public:
         auto attrs = ConvertAttr<MGB::ExternOpr>(op->getAttrDictionary(),
                                                  op->getContext());
         setOperandSegmentAttr(op->getContext(), attrs,
-                              {op.nr_input(), op.nr_output()});
+                              {static_cast<int>(op.nr_input()),
+                               static_cast<int>(op.nr_output())});
         return createOp<Kernel::ExternOpr>(op, operands, rewriter, attrs);
     }
 };
